+++ tensor.h
@@ -19,15 +19,16 @@ public:
     int rows() const;
     int cols() const;
     //运算符
-    float & operator()(int i, int j) const;
+    double & operator()(int i, int j) const;
 	Tensor & operator = (const Tensor& obj);
     //类似于matlab切片操作，不能对切片数组进行赋值操作
 	Tensor slice(int pos, int start_col, int end_col) const;
 	Tensor slice(int start_row, int end_row, int pos) const;
 	Tensor slice(int start_row, int end_row, int start_col, int end_col) const;
+	//non-sense
 	Tensor concatenate(const vector<Tensor> & vec, int axis) const;
 private:
-    float * buffer;
+    double * buffer;
     int row;
     int col;
